repo: plot: convert to package
