+++ tests/func/test_plot.py
@@ -9,16 +9,13 @@ from bs4 import BeautifulSoup
 from funcy import first
 
 from dvc.compat import fspath
+from dvc.repo.plot.data import NoMetricInHistoryError, PlotMetricTypeError
 from dvc.template import (
     DefaultLinearTemplate,
     TemplateNotFound,
     NoDataForTemplateError,
 )
-from dvc.repo.plot import (
-    NoMetricInHistoryError,
-    NoDataOrTemplateProvided,
-    PlotMetricTypeError,
-)
+from dvc.repo.plot import NoDataOrTemplateProvided
 
 
 def _remove_whitespace(value):
